ballbot_rl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ballbot-rl"
version = "0.1.0"
description = "Reinforcement learning training and evaluation for ballbot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ballbot-gym>=0.1.0",  # Local dependency
    "stable-baselines3>=2.0.0",
    "torch>=2.0.0",
    "gymnasium>=0.28.0",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "termcolor>=2.0.0",
]

[project.scripts]
ballbot-train = "ballbot_rl.training.train:cli_main"
ballbot-eval = "ballbot_rl.evaluation.evaluate:cli_main"
ballbot-collect = "ballbot_rl.data.collect:cli_main"
ballbot-pretrain = "ballbot_rl.encoders.pretrain:cli_main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ballbot_rl"]

